refactor(frontend): split git diff panel into workflow and diff-view sections
544-line panel becomes a 173-line container; DiffViewMode alias replaces
a union repeated six times across props; pairDiffLines joins the model.
